Also implement AsUniformValue for Texture2d and friends directly
In addition to implementing if for references, i.e.
`impl<'a> AsUniformValue for &'a {myname}`.

The method `as_uniform_value(&self)` takes self by reference, so there
should be no need to require the second-level reference.

This commit doesn't remove the original impl above in order not to introduce
breaking changes. OTOH it would be cleaner to remove it, thanks to automatic
deref, it should be made redundant by the new impl in a lot of cases.